Rename JSON to jsonV1
