+++ plugins/inputs/zipkin/codec/jsonV1/jsonV1.go
@@ -1,4 +1,4 @@
-package json
+package jsonV1
 
 import (
 	"encoding/json"
